changed structure as lib
