+++ pkg/biwenger/user.go
@@ -1,0 +1,6 @@
+package biwenger
+
+type User struct {
+	Email    string `json:"email"`
+	Password string `json:"password"`
+}
